refactor: Readable CSV arrays of data
