+++ lib/stat_tracker.rb
@@ -1,27 +1,37 @@
 require 'pry'
 require 'csv'
-require './lib/teams'
-require './lib/games'
-require './lib/coaches'
+require_relative './team'
+require_relative './game'
+require_relative './game_team'
 
 class StatTracker
   attr_reader :teams, :games, :game_teams
 
   def initialize(locations)
-    @teams = object_creation(locations[:teams], Team, :team_id)
-    # binding.pry
-    @games = object_creation(locations[:games], Game, :game_id)
-    @game_teams = object_creation(locations[:game_teams], Coach, :game_id)
-    binding.pry
+    @teams = teams
+    @games = games
+    @game_teams = game_teams
   end
 
   def self.from_csv(locations)
-    StatTracker.new(locations)
-  end
+    teams = []
+    teams_data = CSV.read(locations[:teams], headers: true)
+    teams_data.each do |row|
+      teams << Team.new(row)
+    end
+    
+    games = []
+    games_data = CSV.read(locations[:games], headers: true)
+    games_data.each do |row|
+      games << Game.new(row)
+    end
 
-  def object_creation(path, object_type, unique_attr)
-    objects = CSV.read(path, headers: true).map { |row| object_type.new(row) }
-    objects.uniq(&unique_attr)
-    return objects
+    game_teams = []
+    game_teams_data = CSV.read(locations[:game_teams], headers: true)
+    game_teams_data.each do |row|
+      game_teams << GameTeam.new(row)
+    end
+
+    return StatTracker.new(teams, games, game_teams)
   end
 end
