+++ runner.rb
@@ -1,4 +1,4 @@
-require './lib/stat_tracker'
+require_relative './stat_tracker'
 
 game_path = './data/games.csv'
 team_path = './data/teams.csv'
